06.05.2023: Made lines between anchor points

+++ src/main/java/ru/nsu/ccfit/g20202/vartazaryan/utils/Point.java
@@ -7,12 +7,12 @@ public class Point
 {
     @Setter
     @Getter
-    private int x;
+    private double x;
     @Setter
     @Getter
-    private int y;
+    private double y;
 
-    public Point(int x, int y)
+    public Point(double x, double y)
     {
         this.x = x;
         this.y = y;
